fix: creator mode style
